Use JUnit5

+++ legacy/src/test/kotlin/rs/emulate/legacy/archive/ArchiveHashTest.kt
@@ -1,7 +1,8 @@
 package rs.emulate.legacy.archive
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+
 
 /**
  * Tests the [ArchiveUtils.hash] function.
@@ -10,7 +11,7 @@ class ArchiveHashTest {
 
     @Test
     fun test() {
-        val hash = ArchiveUtils.hash(TEST_NAME).toLong()
+        val hash = ArchiveUtils.hash(TEST_NAME)
         assertEquals(EXPECTED, hash)
     }
 
@@ -19,7 +20,7 @@ class ArchiveHashTest {
         /**
          * The expected hash of the [TEST_NAME] string.
          */
-        private const val EXPECTED = 11943852L
+        private const val EXPECTED = 11943852
 
         /**
          * The string to hash.
